Remove buildtags
